Add home story with getting started and list of utilities
