+++ .eslintignore
@@ -1,3 +1,5 @@
+node_modules/
+
 /build/
 /storybook-static/
 /packages/dist-*/
