+++ .storybook/preview.js
@@ -1,3 +1,6 @@
 export const parameters = {
   actions: { argTypesRegex: '^on[A-Z].*' },
+  options: {
+    storySort: { order: ['Home', 'Visualizations'] },
+  },
 };
