+++ src/h5web/visualizations/heatmap/models.ts
@@ -8,8 +8,3 @@ export type Dims = {
 export type D3Interpolator = (t: number) => string;
 
 export type ColorMap = keyof typeof INTERPOLATORS;
-
-export type ColorMapOption = {
-  label: ColorMap;
-  value: ColorMap;
-};
