+++ src/packages/lib.ts
@@ -1,3 +1,48 @@
+// Visualizations
+export { default as ScalarVis } from '../h5web/visualizations/ScalarVis';
+export { default as LineVis } from '../h5web/visualizations/line/LineVis';
 export { default as HeatmapVis } from '../h5web/visualizations/heatmap/HeatmapVis';
+
+// Building blocks
+export { default as AxisSystem } from '../h5web/visualizations/shared/AxisSystem';
+export { default as ColorBar } from '../h5web/visualizations/heatmap/ColorBar';
+
+// Utilities
+export {
+  computeVisSize,
+  findDomain,
+  getSupportedDomain,
+  extendDomain,
+} from '../h5web/visualizations/shared/utils';
+
+export {
+  useDataDomain,
+  useSupportedDomain,
+} from '../h5web/visualizations/shared/hooks';
+
+export {
+  getDims,
+  generateCSSLinearGradient,
+} from '../h5web/visualizations/heatmap/utils';
+
+// Models
+export { ScaleType } from '../h5web/visualizations/shared/models';
+export { CurveType } from '../h5web/visualizations/line/models';
+
+export type {
+  Domain,
+  Size,
+  AxisConfig,
+  IndexAxisConfig,
+  DataAxisConfig,
+  AxisOffsets,
+} from '../h5web/visualizations/shared/models';
+
+export type {
+  Dims,
+  D3Interpolator,
+  ColorMap,
+} from '../h5web/visualizations/heatmap/models';
+
+// Mock data
 export { default as mockData } from '../h5web/providers/mock/data.json';
-export { findDomain } from '../h5web/visualizations/shared/utils';
